Replace TEXTPOS defines with a typed enum
Convert AP settings text positions to a scoped enum and update
callers to use typed values for safer handling.

Bug: 42270078
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/143810
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ fxjs/cjs_field.cpp
@@ -902,7 +902,7 @@ CJS_Result CJS_Field::get_button_position(CJS_Runtime* pRuntime) {
   }
 
   return CJS_Result::Success(
-      pRuntime->NewNumber(pFormControl->GetTextPosition()));
+      pRuntime->NewNumber(static_cast<int>(pFormControl->GetTextPosition())));
 }
 
 CJS_Result CJS_Field::set_button_position(CJS_Runtime* pRuntime,
